Edit how johari calculates frequencies
Previously the johari calculated frequencies such that any adjective in
any window was given a frequency of occurrences in that window divided
by total number of descriptions for the user.

This commit changes unknown and facade adjectives to always have a
frequency of 1 in their windows. Adjectives that no one selected
(unknown) and adjectives you only selected for yourself (facade), in my
opinion, do not really make sense to have frequencies attached with
them.

For arena and blind spot adjectives, frequencies are calculated as
total number of selections by a peer divided by total number of times
the user has been described by a peer. Only peers are included since it
is clear whether you selected it for yourself or not based on its
quadrant.